% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{spaMTPDataResources}
\alias{spaMTPDataResources}
\title{List SpaMTP experiment resources}
\usage{
spaMTPDataResources(version = NULL, category = NULL)
}
\arguments{
\item{version}{Data release version. \code{NULL} returns all versions.}

\item{category}{Optional resource category.}
}
\value{
A data frame describing the registered resources.
}
\description{
List SpaMTP experiment resources
}
\examples{
spaMTPDataResources()
}
