% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{spaMTPDataVersion}
\alias{spaMTPDataVersion}
\title{Report the available SpaMTPData releases}
\usage{
spaMTPDataVersion()
}
\value{
A character vector of data releases, newest first.
}
\description{
Report the available SpaMTPData releases
}
\examples{
spaMTPDataVersion()
}
