% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaMTPData-package.R
\docType{package}
\name{SpaMTPData-package}
\alias{SpaMTPData-package}
\title{SpaMTPData: ExperimentHub data for SpaMTP}
\description{
\code{SpaMTPData} provides discoverable, cached access to the experiment data
used in SpaMTP examples, tests and vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GenomicsMachineLearning/SpaMTPData}
  \item Report bugs at \url{https://github.com/GenomicsMachineLearning/SpaMTPData/issues}
}

}
\author{
\strong{Maintainer}: Tianyao Lu \email{lu.t@wehi.edu.au}

Authors:
\itemize{
  \item Tianyao Lu \email{lu.t@wehi.edu.au}
  \item Andrew Causer
  \item Quan Nguyen
}

}
\keyword{internal}
