% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{SpaMTPDataVersion}
\alias{SpaMTPDataVersion}
\title{Report the available SpaMTPData releases}
\usage{
SpaMTPDataVersion()
}
\value{
A character vector of data releases, newest first.
}
\description{
Report the available SpaMTPData releases
}
\examples{
SpaMTPDataVersion()
}
